+++ test/test_types.py
@@ -22,7 +22,9 @@ def test_none():
 
     test = Test(val=None)
     assert test.val is None
-    test = Test(**{"val": None})
+    assert repr(test) == "Test(val=None)"
+    assert serialize(test) == {"val": None}
+    test = Test(**{'val': None})
     assert test.val is None
 
     with pytest.raises(ShiftError):
@@ -36,6 +38,8 @@ def test_missing_to_none():
 
     test = Test()
     assert test.val is None
+    assert repr(test) == "Test(val=None)"
+    assert serialize(test) == {"val": None}
     test = Test(**{})
     assert test.val is None
 
@@ -50,8 +54,8 @@ def test_int():
 
     test = Test(val=42)
     assert test.val == 42
-    test = Test(**{"val": 42})
-    assert test.val == 42
+    assert repr(test) == "Test(val=42)"
+    assert serialize(test) == {"val": 42}
 
     with pytest.raises(ShiftError):
         _ = Test(val=InvalidType)
@@ -64,7 +68,9 @@ def test_bool():
 
     test = Test(val=True)
     assert test.val is True
-    test = Test(**{"val": True})
+    assert repr(test) == "Test(val=True)"
+    assert serialize(test) == {"val": True}
+    test = Test(**{'val': True})
     assert test.val is True
 
     with pytest.raises(ShiftError):
@@ -78,6 +84,8 @@ def test_float():
 
     test = Test(val=1.21)
     assert test.val == 1.21
+    assert repr(test) == "Test(val=1.21)"
+    assert serialize(test) == {"val": 1.21}
     test = Test(**{"val": 1.21})
     assert test.val == 1.21
 
@@ -92,6 +100,8 @@ def test_str():
 
     test = Test(val="hello there")
     assert test.val == "hello there"
+    assert repr(test) == "Test(val='hello there')"
+    assert serialize(test) == {"val": "hello there"}
     test = Test(**{"val": "hello there"})
     assert test.val == "hello there"
 
@@ -106,6 +116,8 @@ def test_bytes():
 
     test = Test(val=b"hello there")
     assert test.val == b"hello there"
+    assert repr(test) == "Test(val=b'hello there')"
+    assert serialize(test) == {"val": b"hello there"}
     test = Test(**{"val": b"hello there"})
     assert test.val == b"hello there"
 
@@ -120,6 +132,8 @@ def test_bytearray():
 
     test = Test(val=bytearray(b"hello there"))
     assert test.val == bytearray(b"hello there")
+    assert repr(test) == "Test(val=bytearray(b'hello there'))"
+    assert serialize(test) == {"val": bytearray(b"hello there")}
     test = Test(**{"val": bytearray(b"hello there")})
     assert test.val == bytearray(b"hello there")
 
@@ -134,6 +148,8 @@ def test_any():
 
     test = Test(val=42)
     assert test.val == 42
+    assert repr(test) == "Test(val=42)"
+    assert serialize(test) == {"val": 42}
     test = Test(**{"val": 42})
     assert test.val == 42
 
@@ -145,6 +161,8 @@ def test_list():
 
     test = Test(val=[4, 5, 6, 1, 2, 3])
     assert test.val == [4, 5, 6, 1, 2, 3]
+    assert repr(test) == "Test(val=[4, 5, 6, 1, 2, 3])"
+    assert serialize(test) == {"val": [4, 5, 6, 1, 2, 3]}
     test = Test(**{"val": [4, 5, 6, 1, 2, 3]})
     assert test.val == [4, 5, 6, 1, 2, 3]
 
@@ -164,6 +182,14 @@ def test_set():
 
     test = Test(val={4, 5, 6, 1, 2, 3})
     assert test.val == {4, 5, 6, 1, 2, 3}
+    r = repr(test)
+    assert '4' in r
+    assert '5' in r
+    assert '6' in r
+    assert '1' in r
+    assert '2' in r
+    assert '3' in r
+    assert serialize(test) == {"val": {4, 5, 6, 1, 2, 3}}
     test = Test(**{"val": {4, 5, 6, 1, 2, 3}})
     assert test.val == {4, 5, 6, 1, 2, 3}
 
@@ -183,6 +209,15 @@ def test_frozenset():
 
     test = Test(val=frozenset({4, 5, 6, 1, 2, 3}))
     assert test.val == frozenset({4, 5, 6, 1, 2, 3})
+    r = repr(test)
+    assert 'frozenset' in r
+    assert '4' in r
+    assert '5' in r
+    assert '6' in r
+    assert '1' in r
+    assert '2' in r
+    assert '3' in r
+    assert serialize(test) == {"val": frozenset({4, 5, 6, 1, 2, 3})}
     test = Test(**{"val": frozenset({4, 5, 6, 1, 2, 3})})
     assert test.val == frozenset({4, 5, 6, 1, 2, 3})
 
@@ -202,6 +237,8 @@ def test_tuple():
 
     test = Test(val=("hello there", 42))
     assert test.val == ("hello there", 42)
+    assert repr(test) == "Test(val=('hello there', 42))"
+    assert serialize(test) == {"val": ("hello there", 42)}
     test = Test(**{"val": ("hello there", 42)})
     assert test.val == ("hello there", 42)
 
@@ -215,35 +252,35 @@ def test_tuple():
     with pytest.raises(ShiftError):
         _ = Test(**{"val": ("hello there", InvalidType)})
 
-# def test_callable():
-#     class Test(ShiftModel):
-#         val: Callable[[int], str]
-#     @staticmethod
-#     def func(x: int) -> str: return str(x)
-#
-#     test = Test(val=func)
-#     assert test.val(42) == "42"
-#     test = Test(**{"val": func})
-#     assert test.val(42) == "42"
-#
-#     with pytest.raises(ShiftError):
-#         _ = Test(val=InvalidType)
-#     with pytest.raises(ShiftError):
-#         _ = Test(**{"val": InvalidType})
-#
-#     @staticmethod
-#     def func(y: str) -> str: return y
-#     with pytest.raises(ShiftError):
-#         _ = Test(val=func)
-#     with pytest.raises(ShiftError):
-#         _ = Test(**{"val": func})
-#
-#     @staticmethod
-#     def func(x: int) -> int: return x
-#     with pytest.raises(ShiftError):
-#         _ = Test(val=func)
-#     with pytest.raises(ShiftError):
-#         _ = Test(**{"val": func})
+def test_callable():
+    class Test(ShiftModel):
+        val: Callable[[int], str]
+    @staticmethod # noqa
+    def func(x: int) -> str: return str(x)
+
+    test = Test(val=func)
+    assert test.val(42) == "42"
+    test = Test(**{"val": func})
+    assert test.val(42) == "42"
+
+    with pytest.raises(ShiftError):
+        _ = Test(val=InvalidType)
+    with pytest.raises(ShiftError):
+        _ = Test(**{"val": InvalidType})
+
+    @staticmethod # noqa
+    def func(y: str) -> str: return y
+    with pytest.raises(ShiftError):
+        _ = Test(val=func)
+    with pytest.raises(ShiftError):
+        _ = Test(**{"val": func})
+
+    @staticmethod # noqa
+    def func(x: int) -> int: return x
+    with pytest.raises(ShiftError):
+        _ = Test(val=func)
+    with pytest.raises(ShiftError):
+        _ = Test(**{"val": func})
 
 def test_dict():
     class Test(ShiftModel):
@@ -251,6 +288,8 @@ def test_dict():
 
     test = Test(val={"hello there": 42})
     assert test.val == {"hello there": 42}
+    assert repr(test) == "Test(val={'hello there': 42})"
+    assert serialize(test) == {"val": {"hello there": 42}}
     test = Test(**{"val": {"hello there": 42}})
     assert test.val == {"hello there": 42}
 
@@ -270,11 +309,15 @@ def test_union():
 
     test = Test(val=42)
     assert test.val == 42
+    assert repr(test) == "Test(val=42)"
+    assert serialize(test) == {"val": 42}
     test = Test(**{"val": 42})
     assert test.val == 42
 
     test = Test(val="hello there")
     assert test.val == "hello there"
+    assert repr(test) == "Test(val='hello there')"
+    assert serialize(test) == {"val": "hello there"}
     test = Test(**{"val": "hello there"})
     assert test.val == "hello there"
 
@@ -289,10 +332,16 @@ def test_optional():
 
     test = Test(val=42)
     assert test.val == 42
+    assert repr(test) == "Test(val=42)"
+    assert serialize(test) == {"val": 42}
     test = Test(**{"val": 42})
     assert test.val == 42
 
     test = Test()
+    assert test.val is None
+    assert repr(test) == "Test(val=None)"
+    assert serialize(test) == {"val": None}
+    test = Test(**{})
     assert test.val is None
 
 def test_literal():
@@ -301,11 +350,15 @@ def test_literal():
 
     test = Test(val="hello there")
     assert test.val == "hello there"
+    assert repr(test) == "Test(val='hello there')"
+    assert serialize(test) == {"val": "hello there"}
     test = Test(**{"val": "hello there"})
     assert test.val == "hello there"
 
     test = Test(val="I have a bad feeling about this")
     assert test.val == "I have a bad feeling about this"
+    assert repr(test) == "Test(val='I have a bad feeling about this')"
+    assert serialize(test) == {"val": "I have a bad feeling about this"}
     test = Test(**{"val": "I have a bad feeling about this"})
     assert test.val == "I have a bad feeling about this"
 
@@ -319,9 +372,6 @@ def test_literal():
     with pytest.raises(ShiftError):
         _ = Test(**{"val": "invalid"})
 
-class TForwardRef(ShiftModel):
-    val: Optional[ForwardRef("TForwardRef")]
-
 def test_shift():
     class A(ShiftModel):
         val: int
@@ -330,6 +380,8 @@ def test_shift():
 
     test = B(ref=A(val=42))
     assert test.ref.val == 42
+    assert repr(test) == "B(ref=A(val=42))"
+    assert serialize(test) == {"ref": {"val": 42}}
     test = B(**{"ref": A(val=42)})
     assert test.ref.val == 42
 
@@ -343,17 +395,25 @@ def test_shift():
     with pytest.raises(ShiftError):
         _ = B(**{"ref": A(val=InvalidType)})
 
+class TForwardRef(ShiftModel):
+    standard: Optional["TForwardRef"]
+    formal: Optional[ForwardRef("TForwardRef")]
+
 def test_forwardref():
     ref = TForwardRef()
-    test = TForwardRef(val=ref)
-    assert test.val == ref
-    test = TForwardRef(**{"val": ref})
-    assert test.val == ref
+    test = TForwardRef(standard=ref, formal=ref)
+    assert test.standard == ref
+    assert test.formal == ref
+    assert repr(test) == 'TForwardRef(standard=TForwardRef(standard=None, formal=None), formal=TForwardRef(standard=None, formal=None))'
+    assert serialize(test) == {'standard': {'standard': None, 'formal': None}, 'formal': {'standard': None, 'formal': None}}
+
+    test = TForwardRef(**{"standard": ref})
+    assert test.standard == ref
 
     with pytest.raises(ShiftError):
-        _ = TForwardRef(val=InvalidType)
+        _ = TForwardRef(standard=InvalidType)
     with pytest.raises(ShiftError):
-        _ = TForwardRef(**{"val": InvalidType})
+        _ = TForwardRef(**{"standard": InvalidType})
 
 def test_shift_field_basic():
     class Test(ShiftModel):
@@ -361,12 +421,12 @@ def test_shift_field_basic():
 
     test = Test(val=42)
     assert test.val == 42
-
+    assert repr(test) == "Test(val=42)"
+    assert serialize(test) == {"val": 42}
     test = Test(**{"val": 42})
     assert test.val == 42
 
     with pytest.raises(ShiftError):
         _ = Test(val=InvalidType)
-
     with pytest.raises(ShiftError):
         _ = Test(**{"val": InvalidType})